- change default value of –ils to 1 - remove option –iom=nelder_mead_ex2 (use –iom=nelder_mead_ex [–ils=1] instead)
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@16343 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/c/simulation/solver/initialization/method_nelderMeadEx.h
@@ -44,7 +44,7 @@ extern "C"
 {
 #endif
 
-  extern int nelderMeadEx_initialization(INIT_DATA *initData, double *lambdaStart, int lambda_steps);
+  extern int nelderMeadEx_initialization(INIT_DATA *initData, double *lambdaStart, long lambda_steps);
 
 #ifdef __cplusplus
 }
